requirements.txt
pandas==2.1.1
python-dotenv==1.0.0
twilio==8.10.0
tqdm==4.66.1